refact: Large refactor of directory structure and namespacing
