+++ database/factories/ProductImageFactory.php
@@ -1,8 +1,8 @@
 <?php
 
-use IndieHD\Velkart\ProductImage\ProductImage;
 use Illuminate\Http\UploadedFile;
-use IndieHD\Velkart\Product\Product;
+use IndieHD\Velkart\Models\Eloquent\Product;
+use IndieHD\Velkart\Models\Eloquent\ProductImage;
 
 $factory->define(ProductImage::class, function (Faker\Generator $faker) {
