--- src/Attribute/Attribute.php
@@ -1,10 +1,0 @@
-<?php
-
-namespace IndieHD\Velkart\Attribute;
-
-use Illuminate\Database\Eloquent\Model;
-
-class Attribute extends Model
-{
-    public $timestamps = false;
-}
