--- src/Attribute/Contracts/AttributeRepositoryContract.php
@@ -1,8 +1,0 @@
-<?php
-
-namespace IndieHD\Velkart\Attribute\Contracts;
-
-interface AttributeRepositoryContract
-{
-    //
-}
