--- src/Attribute/Repositories/AttributeRepository.php
@@ -1,8 +1,0 @@
-<?php
-
-namespace IndieHD\Velkart\Attribute\Repositories;
-
-class AttributeRepository
-{
-    //
-}
