--- src/AttributeValue/AttributeValue.php
@@ -1,10 +1,0 @@
-<?php
-
-namespace IndieHD\Velkart\AttributeValue;
-
-use Illuminate\Database\Eloquent\Model;
-
-class AttributeValue extends Model
-{
-    public $timestamps = false;
-}
