--- src/AttributeValue/Contracts/AttributeValueRepositoryContract.php
@@ -1,8 +1,0 @@
-<?php
-
-namespace IndieHD\Velkart\AttributeValue\Contracts;
-
-interface AttributeValueRepositoryContract
-{
-    //
-}
