--- src/AttributeValue/Repositories/AttributeValueRepository.php
@@ -1,8 +1,0 @@
-<?php
-
-namespace IndieHD\Velkart\AttributeValue\Repositories;
-
-class AttributeValueRepository
-{
-    //
-}
