--- src/Category/Category.php
@@ -1,12 +1,0 @@
-<?php
-
-namespace IndieHD\Velkart\Category;
-
-use Illuminate\Database\Eloquent\Model;
-
-class Category extends Model
-{
-    public $timestamps = false;
-
-    protected $guarded = ['id'];
-}
