--- src/Category/Contracts/CategoryRepositoryContract.php
@@ -1,8 +1,0 @@
-<?php
-
-namespace IndieHD\Velkart\Category\Contracts;
-
-interface CategoryRepositoryContract
-{
-    //
-}
