--- src/Category/Repositories/CategoryRepository.php
@@ -1,8 +1,0 @@
-<?php
-
-namespace IndieHD\Velkart\Category\Repositories;
-
-class CategoryRepository
-{
-    //
-}
