+++ src/Contracts/CategoryRepositoryContract.php
@@ -1,0 +1,8 @@
+<?php
+
+namespace IndieHD\Velkart\Contracts;
+
+interface CategoryRepositoryContract
+{
+    //
+}
