+++ src/Contracts/ProductRepositoryContract.php
@@ -1,0 +1,8 @@
+<?php
+
+namespace IndieHD\Velkart\Contracts;
+
+interface ProductRepositoryContract extends BaseRepositoryContract
+{
+
+}
