+++ src/Models/Eloquent/Attribute.php
@@ -1,0 +1,10 @@
+<?php
+
+namespace IndieHD\Velkart\Models\Eloquent;
+
+use Illuminate\Database\Eloquent\Model;
+
+class Attribute extends Model
+{
+    public $timestamps = false;
+}
