+++ src/Models/Eloquent/Category.php
@@ -1,0 +1,12 @@
+<?php
+
+namespace IndieHD\Velkart\Models\Eloquent;
+
+use Illuminate\Database\Eloquent\Model;
+
+class Category extends Model
+{
+    public $timestamps = false;
+
+    protected $guarded = ['id'];
+}
