--- src/Product/Contracts/ProductRepositoryContract.php
@@ -1,10 +1,0 @@
-<?php
-
-namespace IndieHD\Velkart\Product\Contracts;
-
-use IndieHD\Velkart\Base\Contracts\BaseRepositoryContract;
-
-interface ProductRepositoryContract extends BaseRepositoryContract
-{
-
-}
