--- src/Product/Product.php
@@ -1,41 +1,0 @@
-<?php
-
-namespace IndieHD\Velkart\Product;
-
-use Gloudemans\Shoppingcart\CanBeBought;
-use Gloudemans\Shoppingcart\Contracts\Buyable;
-use Illuminate\Database\Eloquent\Model;
-use IndieHD\Velkart\Attribute\Attribute;
-use IndieHD\Velkart\Category\Category;
-use IndieHD\Velkart\ProductImage\ProductImage;
-
-class Product extends Model implements Buyable
-{
-    use CanBeBought;
-
-    protected $guarded = ['id'];
-
-    /**
-     * @return \Illuminate\Database\Eloquent\Relations\HasMany
-     */
-    public function images()
-    {
-        return $this->hasMany(ProductImage::class);
-    }
-
-    /**
-     * @return \Illuminate\Database\Eloquent\Relations\BelongsToMany
-     */
-    public function categories()
-    {
-        return $this->belongsToMany(Category::class);
-    }
-
-    /**
-     * @return \Illuminate\Database\Eloquent\Relations\BelongsToMany
-     */
-    public function attributes()
-    {
-        return $this->belongsToMany(Attribute::class);
-    }
-}
