--- src/ProductImage/Contracts/ProductImageRepositoryContract.php
@@ -1,10 +1,0 @@
-<?php
-
-namespace IndieHD\Velkart\ProductImage\Contracts;
-
-use IndieHD\Velkart\Base\Contracts\BaseRepositoryContract;
-
-interface ProductImageRepositoryContract extends BaseRepositoryContract
-{
-
-}
