+++ src/Repositories/Eloquent/AttributeValueRepository.php
@@ -1,0 +1,8 @@
+<?php
+
+namespace IndieHD\Velkart\Repositories\Eloquent;
+
+class AttributeValueRepository
+{
+    //
+}
