+++ src/Repositories/Eloquent/CategoryRepository.php
@@ -1,0 +1,8 @@
+<?php
+
+namespace IndieHD\Velkart\Repositories\Eloquent;
+
+class CategoryRepository
+{
+    //
+}
